fixups of styles
