+++ karapace/protobuf/extend_element.py
@@ -1,16 +1,18 @@
 # Ported from square/wire:
 # wire-library/wire-schema/src/commonMain/kotlin/com/squareup/wire/schema/internal/parser/ExtendElement.kt
-
+from dataclasses import dataclass
+from karapace.protobuf.field_element import FieldElement
 from karapace.protobuf.location import Location
 from karapace.protobuf.utils import append_documentation, append_indented
+from typing import List
 
 
+@dataclass
 class ExtendElement:
-    def __init__(self, location: Location, name: str, documentation: str = "", fields: list = None) -> None:
-        self.location = location
-        self.name = name
-        self.documentation = documentation
-        self.fields = fields or []
+    location: Location
+    name: str
+    documentation: str = ""
+    fields: List[FieldElement] = None
 
     def to_schema(self) -> str:
         result = []
@@ -18,8 +20,8 @@ class ExtendElement:
         result.append(f"extend {self.name} {{")
         if self.fields:
             result.append("\n")
-        for field in self.fields:
-            append_indented(result, field.to_schema())
+            for field in self.fields:
+                append_indented(result, field.to_schema())
 
         result.append("}\n")
         return "".join(result)
