+++ karapace/protobuf/group_element.py
@@ -1,28 +1,21 @@
 # Ported from square/wire:
 # wire-library/wire-schema/src/commonMain/kotlin/com/squareup/wire/schema/internal/parser/GroupElement.kt
+from dataclasses import dataclass
 from karapace.protobuf.field import Field
+from karapace.protobuf.field_element import FieldElement
 from karapace.protobuf.location import Location
 from karapace.protobuf.utils import append_documentation, append_indented
-from typing import Union
+from typing import List, Optional
 
 
+@dataclass
 class GroupElement:
-    def __init__(
-        self,
-        label: Union[None, Field.Label],
-        location: Location,
-        name: str,
-        tag: int,
-        documentation: str = "",
-        fields: list = None
-    ) -> None:
-        self.label = label
-        self.location = location
-        self.name = name
-        self.tag = tag
-
-        self.fields = fields or []
-        self.documentation = documentation
+    label: Optional[Field.Label]
+    location: Location
+    name: str
+    tag: int
+    documentation: str = ""
+    fields: List[FieldElement] = None
 
     def to_schema(self) -> str:
         result = []
