+++ karapace/protobuf/option_element.py
@@ -2,13 +2,7 @@
 # wire-library/wire-schema/src/commonMain/kotlin/com/squareup/wire/schema/internal/parser/OptionElement.kt
 
 from enum import Enum
-# from karapace.protobuf.kotlin_wrapper import *
-# from karapace.protobuf.kotlin_wrapper import *
 from karapace.protobuf.utils import append_indented, append_options, try_to_schema
-
-
-class ListOptionElement(list):
-    pass
 
 
 class OptionElement:
